+++ Twileloop.ExpressSecurity/Encoding/Base64Encoder.cs
@@ -1,16 +1,18 @@
-﻿namespace Twileloop.Security.Abstractions.Encoding
+﻿using System;
+
+namespace Twileloop.Security.Encoding
 {
-    public static class UTF8Encoder
+    public static class Base64Encoder
     {
         public static string Encode(string plainText)
         {
             byte[] plainTextBytes = System.Text.Encoding.UTF8.GetBytes(plainText);
-            return System.Text.Encoding.UTF8.GetString(plainTextBytes);
+            return Convert.ToBase64String(plainTextBytes);
         }
 
         public static string Decode(string encodedText)
         {
-            byte[] encodedBytes = System.Text.Encoding.UTF8.GetBytes(encodedText);
+            byte[] encodedBytes = Convert.FromBase64String(encodedText);
             return System.Text.Encoding.UTF8.GetString(encodedBytes);
         }
     }
