+++ Twileloop.ExpressSecurity/Encoding/UTF8Encoder.cs
@@ -1,18 +1,16 @@
-﻿using System;
-
-namespace Twileloop.Security.Abstractions.Encoding
+﻿namespace Twileloop.Security.Encoding
 {
-    public static class Base64Encoder
+    public static class UTF8Encoder
     {
         public static string Encode(string plainText)
         {
             byte[] plainTextBytes = System.Text.Encoding.UTF8.GetBytes(plainText);
-            return Convert.ToBase64String(plainTextBytes);
+            return System.Text.Encoding.UTF8.GetString(plainTextBytes);
         }
 
         public static string Decode(string encodedText)
         {
-            byte[] encodedBytes = Convert.FromBase64String(encodedText);
+            byte[] encodedBytes = System.Text.Encoding.UTF8.GetBytes(encodedText);
             return System.Text.Encoding.UTF8.GetString(encodedBytes);
         }
     }
