+++ Twileloop.ExpressSecurity/Hashing/MD5Algorithm.cs
@@ -1,7 +1,7 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
 
-namespace Twileloop.Security.Abstractions.Hashing
+namespace Twileloop.Security.Hashing
 {
     public static class MD5Algorithm
     {
