+++ Twileloop.ExpressSecurity/Hashing/SHAAlgorithm.cs
@@ -3,7 +3,7 @@ using Org.BouncyCastle.Crypto.Digests;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace Twileloop.Security.Abstractions.Hashing
+namespace Twileloop.Security.Hashing
 {
     public static class SHAAlgorithm
     {
